RM#30564: Sale order : removed the behavior of removing stock move with a status different than draft after clicking on edit sale order button.

+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderServiceSupplychainImpl.java
@@ -135,7 +135,9 @@ public class SaleOrderServiceSupplychainImpl extends SaleOrderServiceImpl
       }
       for (StockMove stockMove : stockMoves) {
         stockMoveService.cancel(stockMove, cancelReason);
-        stockMoveRepository.remove(stockMove);
+        if (stockMove.getStatusSelect().equals(StockMoveRepository.STATUS_DRAFT)) {
+          stockMoveRepository.remove(stockMove);
+        }
       }
     }
     return checkAvailabiltyRequest;
